+++ src/MCPServer/lib/package.py
@@ -17,6 +17,8 @@
 # You should have received a copy of the GNU General Public License
 # along with Archivematica.  If not, see <http://www.gnu.org/licenses/>.
 
+import abc
+import ast
 import collections
 import logging
 import os
@@ -24,7 +26,10 @@ import shutil
 from tempfile import mkdtemp
 from uuid import uuid4
 
-from django.conf import settings as django_settings
+import scandir
+from lxml import etree
+from django.conf import settings
+from django.utils import six
 
 from archivematicaFunctions import unicodeToStr
 from databaseFunctions import auto_close_db
@@ -32,7 +37,6 @@ from executor import Executor
 from job_chain import JobChain
 from main import models
 import storageService as storage_service
-from unit import Transfer
 
 
 logger = logging.getLogger("archivematica.mcp.server")
@@ -56,55 +60,56 @@ StartingPoint = collections.namedtuple("StartingPoint", "watched_dir chain link"
 PACKAGE_TYPE_STARTING_POINTS = {
     "standard": StartingPoint(
         watched_dir=os.path.join(
-            django_settings.WATCH_DIRECTORY, "activeTransfers/standardTransfer"
+            settings.WATCH_DIRECTORY, "activeTransfers/standardTransfer"
         ),
         chain="6953950b-c101-4f4c-a0c3-0cd0684afe5e",
         link="045c43ae-d6cf-44f7-97d6-c8a602748565",
     ),
     "unzipped bag": StartingPoint(
         watched_dir=os.path.join(
-            django_settings.WATCH_DIRECTORY, "activeTransfers/baggitDirectory"
+            settings.WATCH_DIRECTORY, "activeTransfers/baggitDirectory"
         ),
         chain="c75ef451-2040-4511-95ac-3baa0f019b48",
         link="154dd501-a344-45a9-97e3-b30093da35f5",
     ),
     "zipped bag": StartingPoint(
         watched_dir=os.path.join(
-            django_settings.WATCH_DIRECTORY, "activeTransfers/baggitZippedDirectory"
+            settings.WATCH_DIRECTORY, "activeTransfers/baggitZippedDirectory"
         ),
         chain="167dc382-4ab1-4051-8e22-e7f1c1bf3e6f",
         link="3229e01f-adf3-4294-85f7-4acb01b3fbcf",
     ),
     "dspace": StartingPoint(
-        watched_dir=os.path.join(
-            django_settings.WATCH_DIRECTORY, "activeTransfers/Dspace"
-        ),
+        watched_dir=os.path.join(settings.WATCH_DIRECTORY, "activeTransfers/Dspace"),
         chain="1cb2ef0e-afe8-45b5-8d8f-a1e120f06605",
         link="bda96b35-48c7-44fc-9c9e-d7c5a05016c1",
     ),
     "maildir": StartingPoint(
-        watched_dir=os.path.join(
-            django_settings.WATCH_DIRECTORY, "activeTransfers/maildir"
-        ),
+        watched_dir=os.path.join(settings.WATCH_DIRECTORY, "activeTransfers/maildir"),
         chain="d381cf76-9313-415f-98a1-55c91e4d78e0",
         link="da2d650e-8ce3-4b9a-ac97-8ca4744b019f",
     ),
     "TRIM": StartingPoint(
-        watched_dir=os.path.join(
-            django_settings.WATCH_DIRECTORY, "activeTransfers/TRIM"
-        ),
+        watched_dir=os.path.join(settings.WATCH_DIRECTORY, "activeTransfers/TRIM"),
         chain="e4a59e3e-3dba-4eb5-9cf1-c1fb3ae61fa9",
         link="2483c25a-ade8-4566-a259-c6c37350d0d6",
     ),
     "dataverse": StartingPoint(
         watched_dir=os.path.join(
-            django_settings.WATCH_DIRECTORY, "activeTransfers/dataverseTransfer"
+            settings.WATCH_DIRECTORY, "activeTransfers/dataverseTransfer"
         ),
         # Approve Dataverse Transfer Chain
         chain="10c00bc8-8fc2-419f-b593-cf5518695186",
         # Chain link setting transfer-type: Dataverse
         link="0af6b163-5455-4a76-978b-e35cc9ee445f",
     ),
+}
+
+BASE_REPLACEMENTS = {
+    r"%tmpDirectory%": os.path.join(settings.SHARED_DIRECTORY, "tmp", ""),
+    r"%processingDirectory%": settings.PROCESSING_DIRECTORY,
+    r"%watchDirectoryPath%": settings.WATCH_DIRECTORY,
+    r"%rejectedDirectory%": settings.REJECTED_DIRECTORY,
 }
 
 
@@ -257,7 +262,7 @@ def _move_to_internal_shared_dir(filepath, dest, transfer):
         raise Exception("Error moving from %s to %s (%s)", filepath, dest, e)
     else:
         transfer.currentlocation = dest.replace(
-            django_settings.SHARED_DIRECTORY, "%sharedPath%", 1
+            settings.SHARED_DIRECTORY, "%sharedPath%", 1
         )
         transfer.save()
 
@@ -317,7 +322,7 @@ def create_package(
     @auto_close_db
     def _start(transfer, name, type_, path):
         # TODO: use tempfile.TemporaryDirectory as a context manager in Py3.
-        tmpdir = mkdtemp(dir=os.path.join(django_settings.SHARED_DIRECTORY, "tmp"))
+        tmpdir = mkdtemp(dir=os.path.join(settings.SHARED_DIRECTORY, "tmp"))
         starting_point = PACKAGE_TYPE_STARTING_POINTS.get(type_)
         logger.debug(
             "Package %s: starting transfer (%s)",
@@ -368,7 +373,7 @@ def _determine_transfer_paths(name, path, tmpdir):
         path = os.path.join(path, ".")  # Copy contents of dir but not dir
         transfer_dir = filepath = os.path.join(tmpdir, name)
     return (
-        transfer_dir.replace(django_settings.SHARED_DIRECTORY, "", 1),
+        transfer_dir.replace(settings.SHARED_DIRECTORY, "", 1),
         unicodeToStr(filepath),
         path,
     )
@@ -404,9 +409,7 @@ def _start_package_transfer_with_auto_approval(
     _copy_processing_config(processing_config, transfer.pk, transfer_rel)
 
     logger.debug("Package %s: moving package to processing directory", transfer.pk)
-    _move_to_internal_shared_dir(
-        filepath, django_settings.PROCESSING_DIRECTORY, transfer
-    )
+    _move_to_internal_shared_dir(filepath, settings.PROCESSING_DIRECTORY, transfer)
 
     logger.debug("Package %s: starting workflow processing", transfer.pk)
     unit = Transfer(path, transfer.pk)
@@ -462,13 +465,11 @@ def _copy_processing_config(processing_config, transfer_pk, transfer_rel):
     if processing_config is None:
         return
     src = os.path.join(
-        django_settings.SHARED_DIRECTORY,
+        settings.SHARED_DIRECTORY,
         "sharedMicroServiceTasksConfigs/processingMCPConfigs",
         "%sProcessingMCP.xml" % processing_config,
     )
-    dest = os.path.join(
-        django_settings.SHARED_DIRECTORY, transfer_rel, "processingMCP.xml"
-    )
+    dest = os.path.join(settings.SHARED_DIRECTORY, transfer_rel, "processingMCP.xml")
     logger.debug(
         "Package %s: copying processing configuration" " (from=%s to=%s)",
         transfer_pk,
@@ -512,3 +513,331 @@ class Path(object):
 
     def parts(self):
         return self.uuid, self.path
+
+
+def get_file_replacement_mapping(file_obj, unit_directory):
+    mapping = BASE_REPLACEMENTS.copy()
+    dirname = os.path.dirname(file_obj.currentlocation)
+    name, ext = os.path.splitext(file_obj.currentlocation)
+    name = os.path.basename(name)
+
+    absolute_path = file_obj.currentlocation.replace(r"%SIPDirectory%", unit_directory)
+    absolute_path = absolute_path.replace(r"%transferDirectory%", unit_directory)
+
+    mapping.update(
+        {
+            r"%fileUUID%": file_obj.pk,
+            r"%originalLocation%": file_obj.originallocation,
+            r"%currentLocation%": file_obj.currentlocation,
+            r"%fileGrpUse%": file_obj.filegrpuse,
+            r"%fileDirectory%": dirname,
+            r"%fileName%": name,
+            r"%fileExtension%": ext[1:],
+            r"%fileExtensionWithDot%": ext,
+            r"%relativeLocation%": absolute_path,
+            # TODO: standardize duplicates
+            r"%inputFile%": absolute_path,
+            r"%fileFullName%": absolute_path,
+        }
+    )
+
+    return mapping
+
+
+@six.add_metaclass(abc.ABCMeta)
+class Package(object):
+    """The end result of a workflow (AIP, SIP, etc.)
+    """
+
+    def __init__(self, current_path, uuid):
+        self._current_path = current_path.replace(
+            r"%sharedPath%", settings.SHARED_DIRECTORY
+        )
+        self.uuid = uuid
+
+    def __repr__(self):
+        return '{class_name}("{current_path}", {uuid})'.format(
+            class_name=self.__class__.__name__,
+            uuid=self.uuid,
+            current_path=self.current_path,
+        )
+
+    @property
+    def current_path(self):
+        return self._current_path
+
+    @current_path.setter
+    def current_path(self, value):
+        """Ensure that we always have a real (no shared dir vars) path.
+        """
+        self._current_path = value.replace(r"%sharedPath%", settings.SHARED_DIRECTORY)
+
+    @property
+    def current_path_for_db(self):
+        return self.current_path.replace(settings.SHARED_DIRECTORY, "%sharedPath%", 1)
+
+    @property
+    def package_name(self):
+        basename = os.path.basename(self.current_path.rstrip("/"))
+        return basename.replace("-" + self.uuid, "")
+
+    @property
+    def base_queryset(self):
+        return models.File.objects.filter(sip_id=self.uuid)
+
+    @property
+    def context(self):
+        # This needs to be reloaded from the db every time, because new values
+        # could have been added by a client script.
+        # TODO: pass context changes back from client
+        return PackageContext.load_from_db(self.uuid)
+
+    def reload(self):
+        raise NotImplementedError
+
+    def get_replacement_mapping(self, filter_subdir_path=None):
+        mapping = BASE_REPLACEMENTS.copy()
+        mapping.update(
+            {
+                r"%SIPUUID%": self.uuid,
+                r"%SIPName%": self.package_name,
+                r"%SIPLogsDirectory%": os.path.join(self.current_path, "logs", ""),
+                r"%SIPObjectsDirectory%": os.path.join(
+                    self.current_path, "objects", ""
+                ),
+                r"%SIPDirectory%": self.current_path,
+                r"%SIPDirectoryBasename": os.path.basename(
+                    os.path.abspath(self.current_path)
+                ),
+                r"%relativeLocation%": self.current_path_for_db,
+            }
+        )
+
+        return mapping
+
+    def xmlify(self):
+        ret = etree.Element("unit")
+        etree.SubElement(ret, "type").text = self.__class__.__name__
+        unitXML = etree.SubElement(ret, "unitXML")
+        etree.SubElement(unitXML, "UUID").text = str(self.uuid)
+        etree.SubElement(unitXML, "currentPath").text = self.current_path_for_db
+
+        return ret
+
+    def files(
+        self, filter_filename_start=None, filter_filename_end=None, filter_subdir=None
+    ):
+        """Generator that yields all files associated with the package.
+        """
+        queryset = self.base_queryset
+
+        if filter_filename_start:
+            # TODO: regex filter
+            raise NotImplementedError("filter_filename_start is not implemented")
+        if filter_filename_end:
+            queryset = queryset.filter(currentlocation__endswith=filter_filename_end)
+        if filter_subdir:
+            filter_path = "".join([self.REPLACEMENT_PATH_STRING, filter_subdir])
+            queryset = queryset.filter(currentlocation__startswith=filter_path)
+
+        # If we don't have any matching files in the database, we're in the process of
+        # generating file UUIDs, so walk the filesystem.
+        # TODO: restructure workflow to remove these cases.
+        if not queryset.exists():
+            start_path = self.current_path.encode("utf-8")  # use bytes to return bytes
+            if filter_subdir:
+                start_path = start_path + filter_subdir.encode("utf-8")
+
+            for basedir, subdirs, files in scandir.walk(start_path):
+                for file_name in files:
+                    if (
+                        filter_filename_start
+                        and not file_name.startswith(filter_filename_start)
+                    ) or (
+                        filter_filename_end
+                        and not file_name.endswith(filter_filename_end)
+                    ):
+                        continue
+
+                    file_path = os.path.join(basedir, file_name)
+                    yield {
+                        r"%relativeLocation%": file_path,
+                        r"%fileUUID%": "",
+                        r"%fileGrpUse%": "",
+                    }
+        else:
+            for file_obj in queryset.iterator():
+                yield get_file_replacement_mapping(file_obj, self.current_path)
+
+    def set_variable(self, key, value, chain_link_id):
+        """Sets a UnitVariable, which tracks choices made by users during processing.
+        """
+        # TODO: refactor this concept
+        if not value:
+            value = ""
+        else:
+            value = six.text_type(value)
+
+        unit_var, created = models.UnitVariable.objects.update_or_create(
+            unittype=self.UNIT_VARIABLE_TYPE,
+            unituuid=self.uuid,
+            variable=key,
+            defaults=dict(variablevalue=value, microservicechainlink=chain_link_id),
+        )
+        if created:
+            message = "New UnitVariable %s created for %s: %s (MSCL: %s)"
+        else:
+            message = "Existing UnitVariable %s for %s updated to %s (MSCL" " %s)"
+
+        logger.info(message, key, self.uuid, value, chain_link_id)
+
+
+class DIP(Package):
+    REPLACEMENT_PATH_STRING = r"%SIPDirectory%"
+    UNIT_VARIABLE_TYPE = "DIP"
+    JOB_UNIT_TYPE = "unitDIP"
+
+    def reload(self):
+        """Reload is called for all unit types, but is a no-op for DIPs.
+        """
+        pass
+
+    def get_replacement_mapping(self, filter_subdir_path=None):
+        mapping = super(DIP, self).get_replacement_mapping(
+            filter_subdir_path=filter_subdir_path
+        )
+        mapping[r"%unitType%"] = "DIP"
+
+        if filter_subdir_path:
+            relative_location = filter_subdir_path.replace(
+                settings.SHARED_DIRECTORY, "%sharedPath%", 1
+            )
+            mapping[r"%relativeLocation%"] = relative_location
+
+        return mapping
+
+
+class Transfer(Package):
+    REPLACEMENT_PATH_STRING = r"%transferDirectory%"
+    UNIT_VARIABLE_TYPE = "Transfer"
+    JOB_UNIT_TYPE = "unitTransfer"
+
+    @property
+    def base_queryset(self):
+        return models.File.objects.filter(transfer_id=self.uuid)
+
+    def reload(self):
+        transfer = models.Transfer.objects.get(uuid=self.uuid)
+        self.current_path = transfer.currentlocation
+
+    def get_replacement_mapping(self, filter_subdir_path=None):
+        mapping = super(Transfer, self).get_replacement_mapping(
+            filter_subdir_path=filter_subdir_path
+        )
+
+        mapping.update(
+            {self.REPLACEMENT_PATH_STRING: self.current_path, r"%unitType%": "Transfer"}
+        )
+
+        return mapping
+
+
+class SIP(Package):
+    REPLACEMENT_PATH_STRING = r"%SIPDirectory%"
+    UNIT_VARIABLE_TYPE = "SIP"
+    JOB_UNIT_TYPE = "unitSIP"
+
+    def __init__(self, *args, **kwargs):
+        super(SIP, self).__init__(*args, **kwargs)
+
+        self.aip_filename = None
+        self.sip_type = None
+
+    def reload(self):
+        sip = models.SIP.objects.get(uuid=self.uuid)
+        self.current_path = sip.currentpath
+        self.aip_filename = sip.aip_filename or ""
+        self.sip_type = sip.sip_type
+
+    def get_replacement_mapping(self, filter_subdir_path=None):
+        mapping = super(SIP, self).get_replacement_mapping(
+            filter_subdir_path=filter_subdir_path
+        )
+
+        mapping.update(
+            {
+                r"%unitType%": "SIP",
+                r"%AIPFilename%": self.aip_filename,
+                r"%SIPType%": self.sip_type,
+            }
+        )
+
+        return mapping
+
+
+class PackageContext(object):
+    """Package context tracks choices made previously while processing
+    """
+
+    def __init__(self, *items):
+        self._data = collections.OrderedDict()
+        for key, value in items:
+            self._data[key] = value
+
+    def __repr__(self):
+        return "PackageContext({!r})".format(dict(self._data.items()))
+
+    def __iter__(self):
+        for key, value in six.iteritems(self._data):
+            yield key, value
+
+    def __len__(self):
+        return len(self._data)
+
+    def __contains__(self, key):
+        return key in self._data
+
+    def __getitem__(self, key):
+        return self._data[key]
+
+    def __setitem__(self, key, value):
+        self._data[key] = value
+
+    def __delitem(self, key):
+        del self._data[key]
+
+    @classmethod
+    def load_from_db(cls, uuid):
+        """
+        Loads a context from the UnitVariable table.
+        """
+        context = cls()
+
+        # TODO: we shouldn't need one UnitVariable per chain, with all the same values
+        unit_vars_queryset = models.UnitVariable.objects.filter(
+            unituuid=uuid, variable="replacementDict"
+        )
+        # Distinct helps here, at least
+        unit_vars_queryset = unit_vars_queryset.values_list("variablevalue").distinct()
+        for unit_var_value in unit_vars_queryset:
+            # TODO: nope nope nope, fix eval usage
+            try:
+                unit_var = ast.literal_eval(unit_var_value[0])
+            except (ValueError, SyntaxError):
+                logger.exception(
+                    "Failed to eval unit variable value %s", unit_var_value[0]
+                )
+            else:
+                context.update(unit_var)
+
+        return context
+
+    def copy(self):
+        clone = PackageContext()
+        clone._data = self._data.copy()
+
+        return clone
+
+    def update(self, mapping):
+        for key, value in mapping.items():
+            self._data[key] = value
